Fix up logic for selecting and inserting in python server functions
